+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # [0.3.2] - Unreleased
 - Added zoom methods to gl-map.
+- Made gl-form-page and gl-draw-toolbar responsive.
 
 # [0.3.1] - 2018-06-08
 - Prevent form submission if there are no fields in the page.